A draft of more fine-grained control over logging
We replace the pgaudit.enabled "big red switch" with a pgaudit.log
setting that takes a list of tokens to identify what classes of
commands to log, e.g. «pgaudit.log = "read, write, user"».

We use check and assign hooks on pgaudit.log to convert the tokens into
a bitmap for internal use.

Next, log_audit_event needs to be taught to pay attention to the bitmap
of LogClass values; to do that, we'll also need to classify commands in
the ProcessUtility_hook, and introduce something other than the command
tag to record the classification.

I'm not completely satisfied with the classification right now, but that
can easily be changed once the infrastructure is in place.

Untested proof of concept.