Remove the "if already enabled" check from assign_pgaudit_log
The idea was to prevent audit logging from ever being disabled, but it
didn't work. If pgaudit.log was set to something other than 'none' in
postgresql.conf and some user had been altered to set it to something
different, the latter assignment would be suppressed.

Instead, we'll have to play games with the GucSource in the check hook
to achieve a comparable effect. (I hope.)