Implement proper event filtering
Instead of passing a made up EVENT_xxx value to log_audit_event(), we
now pass in the nodeTag(parsetree) and use that to compare commands
against the configuration in should_be_logged().

Everything else falls out of this change in a pleasantly straightforward
manner. One mild wrinkle is that we don't have access to a parsetree in
the ExecutorCheckPerms_hook, so we have to make up the NodeTag.

The classification of commands into the LogClass categories is
provisional and should be reviewed ASAP.

Untested.