Remove GUC source checking altogether
Nobody seemed to like the idea of even trying to prevent the superuser
from doing anything, and this test provided cosmetic security at best.

But the real problem is that I overlooked the behaviour of PGC_S_TEST
when I wrote the code, and we've received two complaints about "ALTER
DATABASE x SET pgaudit.log …" not working as a result.

Rather than allowing PGC_S_TEST separately and documenting why that
makes sense, it's easier to just remove the test.